requirements.txt
hypothesis
aiohttp
django
pytest
pytest-aiohttp
lxml
channels
asgi_redis
channels_redis
asgi_redis
social-auth-app-django
aiohttp_jinja2
websocket-client

